Handle ULong -> UInt conversions, update api to reflect stream_chacha20_ietf takes UInt for initial counter

+++ multiplatform-crypto-libsodium-bindings/src/jvmMain/kotlin/com/ionspin/kotlin/crypto/JnaLibsodiumInterface.kt
@@ -1236,7 +1236,7 @@ interface JnaLibsodiumInterface : Library {
         message: ByteArray,
         messageLength: Long,
         nonce: ByteArray,
-        initialCounter : Long,
+        initialCounter : Int,
         key: ByteArray
     ) : Int
